+++ src/client/App/style.scss
@@ -65,5 +65,11 @@ body,
   .right {
     flex: 1;
     background-color: white;
+
+    iframe {
+      width: 100%;
+      height: 100%;
+      border: none;
+    }
   }
 }
